[FIX] review users, student, academic year ,,,, and role and permissions

+++ app/Http/Controllers/Backend/CampusBackendController.php
@@ -8,6 +8,7 @@ use App\Contracts\CampusRepositoryInterface;
 use App\Http\Controllers\Controller;
 use App\Http\Requests\CampusRequest;
 use App\Http\Requests\CampusStatusRequest;
+use App\Http\Requests\CampusUpdateRequest;
 use Flasher\SweetAlert\Prime\SweetAlertFactory;
 use Illuminate\Contracts\Foundation\Application;
 use Illuminate\Contracts\Support\Renderable;
@@ -15,20 +16,23 @@ use Illuminate\Contracts\View\Factory;
 use Illuminate\Contracts\View\View;
 use Illuminate\Http\JsonResponse;
 use Illuminate\Http\RedirectResponse;
-use Illuminate\Http\Request;
 use Illuminate\Http\Response as HttpResponse;
+use Illuminate\Support\Facades\Gate;
 use Illuminate\Support\Facades\Response;
+use Symfony\Component\HttpFoundation\Response as SymfonyHttp;
 
 class CampusBackendController extends Controller
 {
     public function __construct(
-        public CampusRepositoryInterface $repository,
-        public SweetAlertFactory $factory
+        protected readonly CampusRepositoryInterface $repository,
+        protected readonly SweetAlertFactory $factory
     ) {
     }
 
     public function index(): Renderable
     {
+        abort_if(Gate::denies('campus-list'), SymfonyHttp::HTTP_FORBIDDEN, '403 Forbidden');
+
         $faculties = $this->repository->getCampuses();
 
         return view('backend.domain.academic.campus.index', compact('faculties'));
@@ -36,9 +40,11 @@ class CampusBackendController extends Controller
 
     public function show(string $key): Factory|View|Application
     {
-        return view('backend.domain.academic.campus.show', [
-            'campus' => $this->repository->showCampus(key:  $key),
-        ]);
+        abort_if(Gate::denies('campus-show'), SymfonyHttp::HTTP_FORBIDDEN, '403 Forbidden');
+
+        $campus = $this->repository->showCampus(key:  $key);
+
+        return view('backend.domain.academic.campus.show', compact('campus'));
     }
 
     public function create(): Renderable
@@ -48,20 +54,24 @@ class CampusBackendController extends Controller
 
     public function store(CampusRequest $attributes): RedirectResponse
     {
+        abort_if(Gate::denies('campus-create'), SymfonyHttp::HTTP_FORBIDDEN, '403 Forbidden');
+
         $this->repository->stored(attributes: $attributes, factory: $this->factory);
 
-        return to_route('admins.academic.campus.index');
+        return redirect()->route('admins.academic.campus.index');
     }
 
     public function edit(string $key): HttpResponse
     {
-        return Response::view('backend.domain.campus.edit', [
-            'campus' => $this->repository->showCampus(key: $key),
-        ]);
+        $campus = $this->repository->showCampus(key:  $key);
+
+        return Response::view('backend.domain.academic.campus.edit', compact('campus'));
     }
 
-    public function update(string $key, CampusRequest $attributes): RedirectResponse
+    public function update(string $key, CampusUpdateRequest $attributes): RedirectResponse
     {
+        abort_if(Gate::denies('campus-edit'), SymfonyHttp::HTTP_FORBIDDEN, '403 Forbidden');
+
         $this->repository->updated(key: $key, attributes: $attributes, factory: $this->factory);
 
         return Response::redirectToRoute('admins.academic.campus.index');
@@ -69,6 +79,8 @@ class CampusBackendController extends Controller
 
     public function destroy(string $key): RedirectResponse
     {
+        abort_if(Gate::denies('campus-delete'), SymfonyHttp::HTTP_FORBIDDEN, '403 Forbidden');
+
         $this->repository->deleted(key: $key, factory: $this->factory);
 
         return Response::redirectToRoute('admins.academic.campus.index');
